% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bind_fadn.R
\name{bind_fadn}
\alias{bind_fadn}
\title{Bind FADN data files together}
\usage{
bind_fadn(directory, country, start_year, end_year, save_as = NULL, save_dir)
}
\arguments{
\item{directory}{String. Path to files.}

\item{country}{String. FADN country code.}

\item{start_year}{Number. First year of the year interval to be combined.}

\item{end_year}{Number. First year of the year interval to be combined.}

\item{save_as}{String. \code{"csv"} to save as a csv file, \code{"R"} to save as Rdata file. If \code{NULL} (defaut) then returns the table.}

\item{save_dir}{String. Saving directory is different than directory.}
}
\description{
For EU FADN data.
}
