% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/decode_vars.R
\name{decode_organic}
\alias{decode_organic}
\title{Decode ORGANIC}
\usage{
decode_organic(data)
}
\arguments{
\item{data}{A dataframe with the following variable: ORGANIC}
}
\value{
data with ORGANIC_c the coded variable and ORGANIC the decoded varaible.
}
\description{
decode the ORGANIC variable
}
