% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/decode_vars.R
\name{decode_SIZ6}
\alias{decode_SIZ6}
\title{Decode SIZ6}
\usage{
decode_SIZ6(data)
}
\arguments{
\item{data}{A dataframe with the following variable: SIZ6}
}
\value{
data with SIZ6_c the coded variable and SIZ6 the decoded varaible.
}
\description{
decode the SIZ6 variable, economic size classes with ES6 EU classification
}
