% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/decode_vars.R
\name{decode_tf14}
\alias{decode_tf14}
\title{Decode TF14}
\usage{
decode_tf14(data)
}
\arguments{
\item{data}{A dataframe with the following variable: TF14}
}
\value{
data with TF14_c the coded variable and TF14 the decoded varaible.
}
\description{
decode the TF14 variable
}
