% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/decode_vars.R
\name{decode_TF8}
\alias{decode_TF8}
\title{Decode TF8}
\usage{
decode_TF8(data)
}
\arguments{
\item{data}{A dataframe with the following variable: TF8}
}
\value{
data with TF8_c the coded variable and TF8 the decoded varaible.
}
\description{
decode the TF8 variable
}
