% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ghg_cresidues.R
\name{ghg_cresidues}
\alias{ghg_cresidues}
\title{Estimate GHG emissions from crop residues}
\usage{
ghg_cresidues(data, crop)
}
\arguments{
\item{data}{A dataframe.}
}
\value{
A dataframe with
\describe{
\item{ghg_cresidues}{GHG emissions from crop residues, expressed in kg CO2e}
}
}
\description{
We use IPCC 2006 Tier 1 estimate (equation 11.6) and emission factors.
}
