% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ghg_elec.R
\name{ghg_elec}
\alias{ghg_elec}
\title{Estimate GHG emissions from electricity consumption for a given crop}
\usage{
ghg_elec(data, crop, emission_factor = "EU")
}
\arguments{
\item{data}{A dataframe. Variables COUNTRY, YEAR, IELE_V must be present.}

\item{crop}{A string - crop code from FADN variables. Must be one of the following: CWHTC, CBRL, CMZ.}

\item{emission_factor}{A string. \code{"country"} when GHG emissions should be estimated using country
specific emission factors. \code{"EU"}, default, when GHG emissions should be estimated using a unique
emission factor for all EU countries.}
}
\value{
A dataframe with
\describe{
\item{ghg_elec}{GHG emissions from electricity inputs expressed in kg CO2e}
}
}
\description{
Emissions from electricity inputs, in kg CO2eq
}
