% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ghg_enteric.R
\name{ghg_enteric}
\alias{ghg_enteric}
\title{Estimate enteric fermentation GHG emissions}
\usage{
ghg_enteric(data)
}
\arguments{
\item{data}{A dataframe.}
}
\value{
A dataframe with 3 columns:
\describe{
\item{Country}{FADN country code}
\item{Year}{Observation year}
\item{GHG_enteric}{GHG emissions from enteric fermentation expressed in tCO2e}
}
}
\description{
CH4 emissions from enteric fermentation. Equation 10.19 from IPCC 2006.
}
