% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ghg_fertilisers.R
\name{ghg_fertilisers}
\alias{ghg_fertilisers}
\title{Estimate GHG emissions from synthetic fertilisers inputs}
\usage{
ghg_fertilisers(data, crop)
}
\arguments{
\item{data}{A dataframe.}

\item{crop}{A string - crop code from FADN variables. Must be one of the following: CWHTC, CBRL, CMZ.}
}
\value{
A dataframe with
\describe{
\item{ghg_ferti_use}{GHG emissions from synthetic fertiliser use, expressed in kg CO2e}
\item{ghg_ferti_prod}{GHG emissions from synthetic fertiliser production, expressed in kg CO2e}
}
}
\description{
We use IPCC 2006 Tier 1 estimate and emission factor.
Direct emissions and emissions from leaching are included.
We chose an economic allocation of fertiliser inputs to each crop production.
}
