% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ghg_fuels.R
\name{ghg_fuels}
\alias{ghg_fuels}
\title{Estimate GHG emissions from fuels consumption for a given crop}
\usage{
ghg_fuels(data, crop, EF_motors = 0.082, EF_heat = 0.0751)
}
\arguments{
\item{data}{A dataframe. Variables COUNTRY, YEAR, IFULS_V and IHFULS_V must be present.}

\item{crop}{A string - crop code from FADN variables. Must be one of the following: CWHTC, CBRL, CMZ.}

\item{EF_motors}{A numeric value. Emission factor for diesel used in motors.
Default value is taken from IPCC 2006, vol 2 - table 3.3.1.}

\item{EF_heat}{A numeric value. Emission factor for diesel, stationary use in agriculture sector.
Default value is taken from IPCC 2006, vol 2 -table 2.5.}
}
\value{
A dataframe with
\describe{
\item{ghg_fuels_motors}{GHG emissions from motor fuels and lubricants expressed in kg CO2e}
\item{ghg_fuels_heat}{GHG emissions from fuels used for heating expressed in kg CO2e}
}
}
\description{
Emissions from motors and heating, in kg CO2eq
}
