% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GHGfromFADN-package.R
\docType{package}
\name{GHGfromFADN-package}
\alias{GHGfromFADN}
\alias{GHGfromFADN-package}
\title{GHGfromFADN: GHG emissions for FADN data}
\description{
GHG emissions from the main productions of the FADN data.
}
\author{
\strong{Maintainer}: Sarah Le Clerc \email{sarah.le-clerc@inrae.fr} (\href{https://orcid.org/0000-0002-0672-5964}{ORCID})

}
\keyword{internal}
