% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zscore.R
\name{plot_score}
\alias{plot_score}
\title{Z-score plot}
\usage{
plot_score(data, variable, threshold = 3, hline = NULL)
}
\arguments{
\item{data}{a dataframe}

\item{variable}{the variable of interest}

\item{threshold}{a number. Threshold for z score. We keep z_i<= threshold}

\item{hline}{a number. Level of the horizontal dash line represented on the plot}
}
\value{
Plot of the z-score with thresholds
}
\description{
z-score is the number of standard deviations from the mean.
Observation i is considered outlier if |z_i|>3
}
