% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict_inuse.R
\name{predict_inuse}
\alias{predict_inuse}
\title{Predict INUSE_Q: quantity of N in mineral fertilisers mineral input}
\usage{
predict_inuse(data, null_org = TRUE)
}
\arguments{
\item{data}{A dataframe with the following variables: COUNTRY, YEAR, ORGANIC, SE025, SE295, SE296 or INUSE_Q}

\item{null_org}{TRUE to set INUSE = 0 for organic and converting farms, default.
FALSE to predict values even for organic and converting farms.}
}
\value{
A dataframe with INUSE_Q_pred expressed in tons and INUSE_Q_pred_kg in kg.
}
\description{
INUSE_Q\_pred is a calculated variable that takes the value of INUSE_Q
from 2017 when available and otherwise a value deducted from the fertiliser
expenditure. By default, INUSE = 0 for organic and converting farms.
}
