% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_fadn.R
\name{read_fadn}
\alias{read_fadn}
\title{Read data}
\usage{
read_fadn(directory, country, year, select_col)
}
\arguments{
\item{directory}{A string. Path to the data files.}

\item{country}{A string. FADN country code.}

\item{year}{A number. Year to read.}

\item{select_col}{A vector or string. Name of columns to select in data, avoids reading the whole - huge - table.}
}
\value{
A dataframe with country, year and columns selected.
}
\description{
Read FADN data
}
