% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{ref_country}
\alias{ref_country}
\title{Country name codes}
\format{
A data frame with 28 rows and 4 variables:
\describe{
\item{country_name}{full country name}
\item{country_code}{country code}
\item{country_fadn}{FADN country code}
\item{country_eu}{Eurostat country code}
}
}
\usage{
ref_country
}
\description{
A dataset containing the code of the EU countries.
}
\keyword{datasets}
