% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shannon_crop.R
\name{shannon_crop}
\alias{shannon_crop}
\title{Shannon crop diversity index}
\usage{
shannon_crop(data)
}
\arguments{
\item{data}{A dataframe. Variables "SE035" "SE041" "SE046" "SE050" "SE054" "SE065" "SE071" "SE075" and "SE025" must be present.}
}
\value{
A dataframe with
\describe{
\item{shannon_cdi}{Shannon Crop Diversity Index for a given farm and year}
}
}
\description{
Calculation of the crop Shannon Diversity Index - SDI.
A higher crop diversity index in absolute value indicates a more adaptative farm -> See Slijper et al 2022.
It should be noted that in this version of the crop diversity index, permanent crops are included.
crops are: cereals, other fieldcrops, vegetables & flowers, vineyards, permanent crops, other permanent crops, forages and woodlands.
}
