% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vtoq_elec.R
\name{vtoq_elec}
\alias{vtoq_elec}
\title{Value to quantity: Electricity IELE_Q}
\usage{
vtoq_elec(data, method = "hh")
}
\arguments{
\item{data}{a dataframe, FADN data. Variables COUNTRY, YEAR and IELE_V must be present.}

\item{method}{a string. Default is "hh" when we take electricity price for households, Eurostat.
Otherwise can be "non_hh" for non-household consumers electricity price.}
}
\value{
data with variable IFULS_Q: Motor fuels and lubricants quantity
}
\description{
Deduction of the quantity in MWH of electricity inputs
from the IELE_V value. We use Eurostat electricity price by country and year, excluding all taxesand levies.
}
