% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vtoq_fuel.R
\name{vtoq_fuel_heating}
\alias{vtoq_fuel_heating}
\title{Value to quantity: Heating fuels IHFULS}
\usage{
vtoq_fuel_heating(data)
}
\arguments{
\item{data}{a dataframe, FADN data. Variable COUNTRY, YEAR and IHFULS_V must be present.}
}
\value{
data with variable IHFULS_Q: Heating fuels quantity
}
\description{
Deduction of the quantity of fuel consumed for heating -in MJ-
from the IHFULS_V value and the price of gasoline for heating from the weekly oil bulletin, eurostat.
}
