% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vtoq_fuel.R
\name{vtoq_fuel_motors}
\alias{vtoq_fuel_motors}
\title{Value to quantity: Motor fuels and lubricants IFULS}
\usage{
vtoq_fuel_motors(data, method = "wob")
}
\arguments{
\item{data}{a dataframe, FADN data. Variables COUNTRY, YEAR and IFULS_V must be present.}

\item{method}{a string. Default is method="wob" to use prices of gasoline excluding taxes
as provided in the weekly oil bulletin - eurostat.
method="unfcc" to deduce a price from the total consumption of fuels in
sector by country and year as provided in the national inventory report - UNFCC.}
}
\value{
data with variable IFULS_Q: Motor fuels and lubricants quantity
}
\description{
Deduction of the quantity of fuel consumed for engines -in MJ-
from the IFULS_V value
}
