% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zscore.R
\name{zscore}
\alias{zscore}
\title{Remove outliers according to z-score}
\usage{
zscore(data, variable, group = NULL, threshold = 3)
}
\arguments{
\item{data}{a dataframe}

\item{variable}{the variable of interest}

\item{group}{a grouping variable. Default is NULL where no grouping variable.}

\item{threshold}{a number. Threshold for z score. We keep z_i<= threshold.}
}
\value{
Data without outliers
}
\description{
z-score is the number of standard deviations from the mean.
Observation i is considered outlier if |z_i|>3
}
